move document type definitions to lib (shared folder)
